improve POr::includes()
Split PConstant and other cases
- A Constant pattern is included if it is included in one of the Or patterns
- An Empty pattern is always included
- Wildcard, Range and Or patterns are included if all their values are included across the Or patterns

The difference is that a Constant pattern needs to be included in only one pattern.
On the other hand Wildcard, Range and Or patterns can have a part included in a first pattern, another part in a second pattern, etc.
This means we need to keep track of the non-included values at each step and compute the difference again.

Results:
- Decreases the number of instances for PEmpty() (136->84) and PatternList() (268->46)
- Might improve performance of check but only visible for small datasets

Misc:
- Change regexr() by ustrregera() in dev_utils.do to capture all tabs